Add functionality and bug fix to DataDisplay class
Add function 'reset' to DataDisplay class that sets the selected_item_ and display_item_ values back to start_
Hopefully temporary solution to bug in DataDisplay class that causes the is_selected_visible() function to return erroneous results.
The bug has to due with the selected_item_ not being included in the clipper. Only seems to occur while animating